+++ tests/basic_tcp.rs
@@ -59,7 +59,7 @@ fn test_sample() {
     config.add_transform(
         "sampler",
         &["in"],
-        transforms::SamplerConfig {
+        transforms::sampler::SamplerConfig {
             rate: 10,
             pass_list: vec![],
         },
@@ -115,14 +115,14 @@ fn test_parse() {
     config.add_transform(
         "parser",
         &["in"],
-        transforms::RegexParserConfig {
+        transforms::regex_parser::RegexParserConfig {
             regex: r"status=(?P<status>\d+)".to_string(),
         },
     );
     config.add_transform(
         "filter",
         &["parser"],
-        transforms::FieldFilterConfig {
+        transforms::field_filter::FieldFilterConfig {
             field: "status".to_string(),
             value: "404".to_string(),
         },
